refactor: move holdings util functions to one util file
