--- src/components/sumTotalHoldings.tsx
@@ -1,6 +1,0 @@
-import { HoldingsApi } from "src/service/holdings";
-import { sumCeloTotal, sumNonCelo } from "./Holdings";
-
-export function sumTotalHoldings(holdings: HoldingsApi) {
-  return sumCeloTotal(holdings) + sumNonCelo(holdings);
-}
